installer: make sure we can always test the installer in ci and locally
Just now there is a dependency on cachix, which means we cannot test
the installer in CI if forks do not have the necessary secrets set up.
We replace this with a simple http server that serves the installer and
can be both used in CI and locally.

(cherry picked from commit 6162e1220e57031c980a3269ed8cc7d34ab46c0b)

# Conflicts:
#	.github/workflows/ci.yml
#	flake.nix